[internal] Port `default_cache_path()` and `all_counter_names()` to PyO3
# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]
