+++ src/python/pants/engine/internals/native_engine.pyi
@@ -46,7 +46,6 @@ def run_local_interactive_process(
 def write_digest(
     scheduler: PyScheduler, session: PySession, digest: PyDigest, path_prefix: str
 ) -> None: ...
-def default_cache_path() -> str: ...
 def write_log(msg: str, level: int, target: str) -> None: ...
 def flush_log() -> None: ...
 def set_per_run_log_path(path: str | None) -> None: ...
@@ -121,7 +120,6 @@ def session_record_test_observation(
     scheduler: PyScheduler, session: PySession, value: int
 ) -> None: ...
 def session_isolated_shallow_clone(session: PySession, build_id: str) -> PySession: ...
-def all_counter_names() -> list[str]: ...
 def graph_len(scheduler: PyScheduler) -> int: ...
 def graph_visualize(scheduler: PyScheduler, session: PySession, path: str) -> None: ...
 def graph_invalidate(scheduler: PyScheduler, paths: Sequence[str]) -> int: ...
