+++ src/python/pants/engine/internals/native_engine_pyo3.pyi
@@ -9,7 +9,29 @@ from pants.engine.fs import PathGlobs
 #   see https://github.com/psf/black/issues/1548
 # flake8: noqa: E302
 
+# ------------------------------------------------------------------------------
+# Scheduler
+# ------------------------------------------------------------------------------
+
+class PyExecutor:
+    def __init__(self, core_threads: int, max_threads: int) -> None: ...
+
+# ------------------------------------------------------------------------------
+# FS
+# ------------------------------------------------------------------------------
+
+def default_cache_path() -> str: ...
 def match_path_globs(path_globs: PathGlobs, paths: tuple[str, ...]) -> str: ...
+
+# ------------------------------------------------------------------------------
+# Workunits
+# ------------------------------------------------------------------------------
+
+def all_counter_names() -> list[str]: ...
+
+# ------------------------------------------------------------------------------
+# Nailgun
+# ------------------------------------------------------------------------------
 
 class PyNailgunClient:
     def __init__(self, port: int, executor: PyExecutor) -> None: ...
@@ -21,8 +43,9 @@ class PantsdConnectionException(Exception):
 class PantsdClientException(Exception):
     pass
 
-class PyExecutor:
-    def __init__(self, core_threads: int, max_threads: int) -> None: ...
+# ------------------------------------------------------------------------------
+# Testutil
+# ------------------------------------------------------------------------------
 
 class PyStubCASBuilder:
     def always_errors(self) -> PyStubCASBuilder: ...
