+++ src/rust/engine/engine_pyo3/src/externs/interface.rs
@@ -7,12 +7,14 @@ use pyo3::prelude::*;
 mod fs;
 mod nailgun;
 mod testutil;
+mod workunits;
 
 #[pymodule]
 fn native_engine_pyo3(py: Python, m: &PyModule) -> PyResult<()> {
   self::fs::register(m)?;
   self::nailgun::register(py, m)?;
   self::testutil::register(m)?;
+  self::workunits::register(m)?;
 
   m.add_class::<PyExecutor>()?;
