+++ src/rust/engine/engine_pyo3/src/externs/interface/workunits.rs
@@ -1,0 +1,15 @@
+// Copyright 2021 Pants project contributors (see CONTRIBUTORS.md).
+// Licensed under the Apache License, Version 2.0 (see LICENSE).
+
+use pyo3::prelude::*;
+use workunit_store::Metric;
+
+pub(crate) fn register(m: &PyModule) -> PyResult<()> {
+  m.add_function(wrap_pyfunction!(all_counter_names, m)?)?;
+  Ok(())
+}
+
+#[pyfunction]
+fn all_counter_names() -> Vec<String> {
+  Metric::all_metrics()
+}
